+++ include/Game/Game.h
@@ -13,7 +13,6 @@
 #include "Input/KeyboardInputHandler.h"
 #include "Input/MouseInputHandler.h"
 #include "Core/EKeyDown.h"
-#include "Core/EntityXMLFile.h"
 #include "Core/EntityFilePreprocessor.h"
 #include "Core/SystemPipeline.h"
 #include "RaptorCopterSystem.h"
